+++ src/reader.rs
@@ -323,107 +323,9 @@ impl FtdcReader {
                         "Processing metric document (type 1) with chunk of binary data: {} bytes",
                         bin.bytes.len()
                     );
-
-                    // Extract numeric values from reference document to use for delta decoding
-                    let mut reference_values = Vec::new();
-                    let mut ref_keys = Vec::new();
-
-                    let mut cursor = Cursor::new(&bin.bytes);
-
-                    // Parse the BSON bytes into a Document
-                    let ref_doc = &Document::from_reader(&mut cursor).map_err(|e| ReaderError::InvalidFile(format!("Failed to parse BSON document {}", e)))?;
-                    fn extract_numeric_values_recursive(
-                        doc: &Document,
-                        prefix: &str,
-                        keys: &mut Vec<String>,
-                        values: &mut Vec<u64>,
-                    ) {
-                        for (key, value) in doc.iter() {
-                            let field_name = if prefix.is_empty() {
-                                key.clone()
-                            } else {
-                                format!("{}_{}", prefix, key)
-                            };
-
-                            match value {
-                                Bson::Double(v) => {
-                                    keys.push(field_name);
-                                    values.push(v.to_bits());
-                                }
-                                Bson::Int32(v) => {
-                                    keys.push(field_name);
-                                    values.push(*v as u64);
-                                }
-                                Bson::Int64(v) => {
-                                    keys.push(field_name);
-                                    values.push(*v as u64);
-                                }
-                                Bson::Boolean(v) => {
-                                    keys.push(field_name);
-                                    values.push(*v as u64);
-                                }
-                                Bson::Document(subdoc) => {
-                                    extract_numeric_values_recursive(
-                                        subdoc,
-                                        &field_name,
-                                        keys,
-                                        values,
-                                    );
-                                }
-                                Bson::Array(arr) => {
-                                    for (i, item) in arr.iter().enumerate() {
-                                        let array_name = format!("{}_{}", field_name, i);
-                                        match item {
-                                            Bson::Document(subdoc) => {
-                                                extract_numeric_values_recursive(
-                                                    subdoc,
-                                                    &array_name,
-                                                    keys,
-                                                    values,
-                                                );
-                                            }
-                                            Bson::Double(v) => {
-                                                keys.push(array_name);
-                                                values.push(v.to_bits());
-                                            }
-                                            Bson::Int32(v) => {
-                                                keys.push(array_name);
-                                                values.push(*v as u64);
-                                            }
-                                            Bson::Int64(v) => {
-                                                keys.push(array_name);
-                                                values.push(*v as u64);
-                                            }
-                                            Bson::Boolean(v) => {
-                                                keys.push(array_name);
-                                                values.push(*v as u64);
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-
-                    extract_numeric_values_recursive(
-                        ref_doc,
-                        "",
-                        &mut ref_keys,
-                        &mut reference_values,
-                    );
-
-                    println!(
-                        "\tReference document has {} key/value pairs",
-                        reference_values.len()
-                    );
-
                     // Decompress using reference values
                     let decompressed = match Compression::decompress_metrics_chunk(
-                        &bin.bytes,
-                        Some(&reference_values),
-                    ) {
+                        &bin.bytes                    ) {
                         Ok(values) => values,
                         Err(e) => {
                             println!("Decompression error: {:?}", e);
